SMTP: Cc and Bcc are recipients too
In SMTP, the RCPT should also receive the Cc: and Bcc: lists of
the E-Mail, otherwise, the mail is only sent to the recipients
in the To: list, because servers don't do the extra work of
inferring the additional recipients from these headers.

Issue discovered while passing an non empty list in Cc and
wondering why the messages did not reach the other side.