Strip carriage returns from the end of lines before sending
This fixes an issue reported on haskell-cafe[1], where spurious
linefeeds will find their way into an email sent with HaskellNet.  The
reason is that we use ByteString's `lines` function to split the source
ByteString into lines, and then join those lines together again with
'\r\n' -- but `lines` only splits on '\n', so if the source ByteString
already used '\r\n'-style linefeeds the resulting mail ends up with
extra carriage returns.

This commit takes the simplest approach to solving this: after splitting
up the lines on '\n', we check if any of the resulting lines end in
'\r', and if they do we strip it.  This covers us for '\n' and '\r\n'
linefeeds, but not for '\r' linefeeds.

[1]: http://article.gmane.org/gmane.comp.lang.haskell.cafe/120647