[issue-33] Simplify mail sending API.
Instead of having mulitple methods with puzzling names for sending
emails, we use only one instead. We provided documentation how
to use mime-mail package to mimic old functions. The change is safe
as all the functions were just a wrappers over the mime-mail functions,
but those wrappers does not provide all the functionality of the package.

Current approach should be much simpler and error prone.