Hide erros about non-exhaustive pattern matches
There is a problem in sendCommand message that we process some
commands first and compiler can't infer that they are not reachable
in the next clauses. It's possible to ignore the message, but it will
require developer to observe the code each time in order to check
if we have not introduced a new error. Another possibility is to
introduce a datatype for all the commands and try make GHC infter
the unreachability of the clause — probably it's the best scenario
but it's a major change to the library, that can affect downstream
packages. So now we added a error message that works as a comment to
the code and will provide a reasonable error to the user if ever
triggered